backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
supabase==2.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
xgboost==1.7.6
lightgbm==4.1.0
catboost==1.2.2
joblib==1.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2